feat: renaming avatar to index, updating snapshots
