+++ app/components/Avatar/default-avatar.png
