+++ app/components/Avatar/index.story.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react'
 import { storiesOf } from '@storybook/react-native'
 import { StoryScreen, Story, UseCase } from '../../../storybook/views'
-import Avatar from './Avatar'
+import Avatar from './'
 
 declare var module
 
 storiesOf('Avatar', module)
   .addDecorator((fn) => <StoryScreen>{fn()}</StoryScreen>)
-  .add('With default avatar', () => (
+  .add('With default Avatar', () => (
     <Story>
       <UseCase text="Uses no thumbnail" usage="While not defining thumbnail, default avatar should display.">
         <Avatar />
