+++ app/components/index.ts
@@ -8,3 +8,4 @@ export * from './switch/switch'
 export * from './text/text'
 export * from './text-field/text-field'
 export * from './wallpaper/wallpaper'
+export { default as Avatar } from './Avatar'
